+++ examples/shared/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vultisig/examples-shared
 
+## 0.1.23
+
+### Patch Changes
+
+- Updated dependencies [[`e434998`](https://github.com/vultisig/vultisig-sdk/commit/e434998069e6af9664db045c5e91c5d5f35feef6)]:
+  - @vultisig/sdk@0.22.6
+
 ## 0.1.22
 
 ### Patch Changes